Fixed size available check for large drives (integer overflow)

+++ src/org/digitalcampus/oppia/utils/storage/FileUtils.java
@@ -271,7 +271,7 @@ public class FileUtils {
             bytesAvailable = stat.getBlockSizeLong() * stat.getAvailableBlocksLong();
         }
         else{
-            bytesAvailable = stat.getBlockSize() * stat.getAvailableBlocks();
+            bytesAvailable = (long)stat.getBlockSize() * (long)stat.getAvailableBlocks();
         }
         return bytesAvailable;
     }
